# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  code: string;
  name: string;
  price: number;
  cost: number;
}

export interface CreateProduct {
  code: string;
  name: string;
  price: number;
  cost: number;
}

export interface Ingredient {
  code: string;
  name: string;
  unit: string;
  stock: number;
  minimum: number;
}

export interface CreateIngredient {
  code: string;
  name: string;
  unit: string;
  stock: number;
  minimum?: number;
}