# lib/api-zod/src/generated/types/createIngredient.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIngredient {
  code: string;
  name: string;
  unit: string;
  stock: number;
  minimum?: number;
}